+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix installation's empty secret scenario.
+
 ## [0.17.0] - 2024-03-12
 
 ### Changed